requirements.txt
Django==1.11
django-extra-fields==1.0.0
djangorestframework==3.8
psycopg2==2.7.3
pytz==2017.2
